requirements.txt
youtube-transcript-api==1.2.4
requests
yt-dlp
praat-parselmouth
numpy
